Adds CORS support to the Fastify server
This change introduces the @fastify/cors package to the project, which allows the server to handle Cross-Origin Resource Sharing (CORS) and thus, enables the secure exchange of data between different domains.

In terms of code changes, the @fastify/cors has been added as a dependency in the package.json and package-lock.json files. Furthermore, the Fastify server now registers @fastify/cors, allowing the server to manage CORS effectively.